requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==1.11.15
django-dotenv==1.4.2
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
iso8601==0.1.12
Markdown==2.6.11
promise==2.1
pycparser==2.18
PyJWT==1.6.4
PyMySQL==0.9.2
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
